extract common code from repls
